+++ packages/gestalt/src/Avatar/InternalAvatar.tsx
@@ -1,9 +1,14 @@
 import { useState } from 'react';
 import classnames from 'classnames';
-import { TOKEN_COLOR_BACKGROUND_AVATAR_PLACEHOLDER } from 'gestalt-design-tokens';
+import {
+  TOKEN_COLOR_BACKGROUND_AVATAR_PLACEHOLDER,
+  TOKEN_COLOR_BACKGROUND_DEFAULT,
+  TOKEN_SPACE_0,
+} from 'gestalt-design-tokens';
 import avatarStyles from './AvatarFoundation.css';
 import DefaultAvatar from './DefaultAvatar';
 import Box from '../Box';
+import Icon from '../Icon';
 import IconCompact from '../IconCompact';
 import Image from '../Image';
 import Mask from '../Mask';
@@ -15,6 +20,15 @@ const sizes = {
   md: 48,
   lg: 64,
   xl: 120,
+} as const;
+
+const verifiedIconSizes = {
+  xs: 14,
+  sm: 14,
+  md: 14,
+  lg: 16,
+  xl: 24,
+  fit: 14,
 } as const;
 
 type Props = {
@@ -50,13 +64,22 @@ function InternalAvatar(props: Props) {
   const width = size === 'fit' ? '100%' : sizes[size];
   const height = size === 'fit' ? '' : sizes[size];
 
+  const verifiedIconPadding = {
+    xs: TOKEN_SPACE_0,
+    sm: TOKEN_SPACE_0,
+    md: TOKEN_SPACE_0,
+    lg: isInVRExperiment ? 'var(--sema-space-50)' : '2px',
+    xl: '5px',
+    fit: TOKEN_SPACE_0,
+  } as const;
+
   return (
     <Box
       dangerouslySetInlineStyle={{
         __style: outline
           ? {
               outline: isInVRExperiment
-                ? '2px solid var(--sema-color-border-inverse)'
+                ? '1px solid var(--sema-color-border-inverse)'
                 : '1px solid rgb(255 255 255)',
             }
           : {},
@@ -97,26 +120,39 @@ function InternalAvatar(props: Props) {
 
       {verified && (
         <Box
+          color="default"
           dangerouslySetInlineStyle={{
-            __style: {
-              bottom: '4%',
-              right: '4%',
-            },
+            __style:
+              size === 'xl'
+                ? {
+                    bottom: verifiedIconPadding[size],
+                    right: verifiedIconPadding[size],
+                    outline: isInVRExperiment
+                      ? '1px solid var(--sema-color-background-default)'
+                      : `1px solid ${TOKEN_COLOR_BACKGROUND_DEFAULT}`,
+                  }
+                : {
+                    bottom: verifiedIconPadding[size],
+                    right: verifiedIconPadding[size],
+                  },
           }}
-          height="25%"
-          minHeight={12}
-          minWidth={12}
+          height={size === 'fit' ? '25%' : verifiedIconSizes[size]}
+          minHeight={size === 'fit' ? verifiedIconSizes[size] : undefined}
+          minWidth={size === 'fit' ? verifiedIconSizes[size] : undefined}
           position="absolute"
-          width="25%"
+          rounding="circle"
+          width={size === 'fit' ? '25%' : verifiedIconSizes[size]}
         >
-          <Box color="default" height="100%" rounding="circle" width="100%">
+          {size === 'xl' ? (
+            <Icon accessibilityLabel="" color="brandPrimary" icon="check-circle-fill" size="100%" />
+          ) : (
             <IconCompact
               accessibilityLabel=""
               color="brandPrimary"
               icon="compact-check-circle-fill"
               size="100%"
             />
-          </Box>
+          )}
         </Box>
       )}
     </Box>