+++ packages/gestalt/src/Tabs.test.tsx
@@ -2,27 +2,6 @@ import { create } from 'react-test-renderer';
 import Tabs from './Tabs';
 
 describe('<Tabs />', () => {
-  // TODO: we no longer support this, but we should
-  // eslint-disable-next-line jest/no-commented-out-tests
-  // test('Only add aria-selected to the active tab', () => {
-  //   const instance = create(
-  //     <Tabs
-  //       tabs={[
-  //         { text: 'News', href: '#' },
-  //         { text: 'You', href: '#' },
-  //         { text: 'Messages', href: '#' },
-  //       ]}
-  //       activeTabIndex={0}
-  //       onChange={() => {}}
-  //     />,
-  //   ).root;
-  //   const tabs = instance.findAllByType(TabWithForwardRef);
-
-  //   expect(tabs[0].props['aria-selected']).toEqual(true);
-  //   expect(tabs[1].props['aria-selected']).toEqual(false);
-  //   expect(tabs[2].props['aria-selected']).toEqual(false);
-  // });
-
   test('Adds id when given', () => {
     const instance = create(
       <Tabs
@@ -44,6 +23,21 @@ describe('<Tabs />', () => {
       <Tabs
         activeTabIndex={0}
         onChange={() => {}}
+        tabs={[
+          { text: 'News', href: '#' },
+          { text: 'You', href: '#' },
+        ]}
+      />,
+    ).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  test('matches snapshot for size lg', () => {
+    const tree = create(
+      <Tabs
+        activeTabIndex={0}
+        onChange={() => {}}
+        size="lg"
         tabs={[
           { text: 'News', href: '#' },
           { text: 'You', href: '#' },