+++ packages/gestalt/src/Tabs.tsx
@@ -26,6 +26,10 @@ type Props = {
     dangerouslyDisableOnNavigation: () => void;
   }) => void;
   /**
+   * See the [size variant](https://gestalt.pinterest.systems/web/tabs#Size) variant to learn more.
+   */
+  size?: 'sm' | 'lg';
+  /**
    * The array of tabs to be displayed. The active tab (as indicated by `activeTabIndex`) will be underlined. Use the optional `indicator` field to show a notification of new items on the tab — see the [indicator variant](https://gestalt.pinterest.systems/web/tabs#Indicator) to learn more. Though `text` currently accepts a React.Node, this is deprecated and will be replaced by a simple `string` type soon.
    */
   tabs: ReadonlyArray<{
@@ -55,6 +59,7 @@ export default function Tabs({
   bgColor = 'default',
   onChange,
   tabs,
+  size = 'sm',
   wrap,
   dataTestId,
 }: Props) {
@@ -83,6 +88,7 @@ export default function Tabs({
           indicator={indicator}
           isActive={activeTabIndex === index}
           onChange={onChange}
+          size={size}
           text={text}
         />
       ))}