+++ CHANGELOG.md
@@ -1,3 +1,81 @@
+## 174.3.1 (Mar 26, 2025)
+
+### Patch
+
+- SearchGuide: (mWeb) update color variant "11" to match figma in dark mode ([#4050](https://github.com/pinterest/gestalt/pull/4050)) - [Preview link](https://deploy-preview-4050--gestalt.netlify.app?devexample=true)
+
+## 174.3.0 (Mar 24, 2025)
+
+### Minor
+
+- AvatarGroupCluster: new component ([#4048](https://github.com/pinterest/gestalt/pull/4048)) - [Preview link](https://deploy-preview-4048--gestalt.netlify.app?devexample=true)
+
+## 174.2.2 (Mar 24, 2025)
+
+### Patch
+
+- Masonry: Fix scroll position glitch on load ([#4045](https://github.com/pinterest/gestalt/pull/4045)) - [Preview link](https://deploy-preview-4045--gestalt.netlify.app?devexample=true)
+
+## 174.2.1 (Mar 21, 2025)
+
+### Patch
+
+- SearchField: fix rtl in VR bug ([#4047](https://github.com/pinterest/gestalt/pull/4047)) - [Preview link](https://deploy-preview-4047--gestalt.netlify.app?devexample=true)
+
+## 174.2.0 (Mar 20, 2025)
+
+### Minor
+
+- Icon: New Icons [GESTALT-8749] ([#4038](https://github.com/pinterest/gestalt/pull/4038)) - [Preview link](https://deploy-preview-4038--gestalt.netlify.app?devexample=true)
+
+## 174.1.2 (Mar 19, 2025)
+
+### Patch
+
+- RadioGroup: fix in CSS file for width ([#4046](https://github.com/pinterest/gestalt/pull/4046)) - [Preview link](https://deploy-preview-4046--gestalt.netlify.app?devexample=true)
+
+## 174.1.1 (Mar 19, 2025)
+
+### Patch
+
+- Internal: [Snyk] Security upgrade react-cookie from 4.1.1 to 8.0.1 ([#4044](https://github.com/pinterest/gestalt/pull/4044)) - [Preview link](https://deploy-preview-4044--gestalt.netlify.app?devexample=true)
+
+## 174.1.0 (Mar 18, 2025)
+
+### Minor
+
+- Tabs: supports large size `size="lg"` ([#4042](https://github.com/pinterest/gestalt/pull/4042)) - [Preview link](https://deploy-preview-4042--gestalt.netlify.app?devexample=true)
+
+## 174.0.0 (Mar 18, 2025)
+
+### Major
+
+- RadioButton: deprecate component ([#4043](https://github.com/pinterest/gestalt/pull/4043)) - [Preview link](https://deploy-preview-4043--gestalt.netlify.app?devexample=true)
+
+## 173.2.2 (Mar 18, 2025)
+
+### Patch
+
+- TileData: removed width and max width definitions ([#4041](https://github.com/pinterest/gestalt/pull/4041)) - [Preview link](https://deploy-preview-4041--gestalt.netlify.app?devexample=true)
+
+## 173.2.1 (Mar 17, 2025)
+
+### Patch
+
+- Avatar: updates to verified icon ([#4040](https://github.com/pinterest/gestalt/pull/4040)) - [Preview link](https://deploy-preview-4040--gestalt.netlify.app?devexample=true)
+
+## 173.2.0 (Mar 17, 2025)
+
+### Minor
+
+- ChartGraph: added support to custom labels in bars + new ChartGraph.Label ([#4039](https://github.com/pinterest/gestalt/pull/4039)) - [Preview link](https://deploy-preview-4039--gestalt.netlify.app?devexample=true)
+
+## 173.1.2 (Mar 17, 2025)
+
+### Patch
+
+- Internal: Bump @babel/runtime from 7.15.4 to 7.26.10 ([#4036](https://github.com/pinterest/gestalt/pull/4036)) - [Preview link](https://deploy-preview-4036--gestalt.netlify.app?devexample=true)
+
 ## 173.1.1 (Mar 10, 2025)
 
 ### Patch